+++ packages/rich-text-with-backend/package.json
@@ -24,19 +24,20 @@
   "dependencies": {
     "@radix-ui/react-popover": "1.1.6",
     "@tippyjs/react": "4.2.6",
-    "@tiptap/core": "2.11.3",
-    "@tiptap/extension-bullet-list": "2.11.3",
-    "@tiptap/extension-code-block": "2.11.3",
-    "@tiptap/extension-code-block-lowlight": "2.11.3",
-    "@tiptap/extension-image": "2.11.3",
-    "@tiptap/extension-link": "2.11.3",
-    "@tiptap/extension-ordered-list": "2.11.3",
-    "@tiptap/extension-paragraph": "2.11.3",
-    "@tiptap/extension-placeholder": "2.11.3",
-    "@tiptap/extension-underline": "2.11.3",
-    "@tiptap/react": "2.11.3",
-    "@tiptap/starter-kit": "2.11.3",
-    "@tiptap/suggestion": "2.11.3",
+    "@tiptap/core": "2.11.5",
+    "@tiptap/extension-bullet-list": "2.11.5",
+    "@tiptap/extension-code-block": "2.11.5",
+    "@tiptap/extension-code-block-lowlight": "2.11.5",
+    "@tiptap/extension-image": "2.11.5",
+    "@tiptap/extension-link": "2.11.5",
+    "@tiptap/extension-ordered-list": "2.11.5",
+    "@tiptap/extension-paragraph": "2.11.5",
+    "@tiptap/extension-placeholder": "2.11.5",
+    "@tiptap/extension-underline": "2.11.5",
+    "@tiptap/extension-heading": "2.11.5",
+    "@tiptap/react": "2.11.5",
+    "@tiptap/starter-kit": "2.11.5",
+    "@tiptap/suggestion": "2.11.5",
     "i18next": "24.2.2",
     "lowlight": "3.3.0",
     "lucide-react": "0.474.0",
