+++ packages/rich-text-with-backend/src/hooks/useTextMenuStates.ts
@@ -1,8 +1,8 @@
 import { Editor } from '@tiptap/react';
 import { useCallback } from 'react';
 import { ShouldShowProps } from '../extensions/menus/TextMenu/types.js';
-import isCustomNodeSelected from '../utilities/isCustomNodeSelected.js';
-import isTextSelected from '../utilities/isTextSelected.js';
+import isCustomNodeSelected from '../lib/utils/isCustomNodeSelected.js';
+import isTextSelected from '../lib/utils/isTextSelected.js';
 
 export const useTextMenuStates = (editor: Editor) => {
   const shouldShow = useCallback(
