+++ packages/rich-text-with-backend/src/lib/utils/isTextSelected.ts
