--- packages/rich-text-with-backend/src/parts/Tooltip/types.ts
@@ -1,17 +1,0 @@
-import React from 'react'
-import { Placement, Props } from 'tippy.js'
-
-export interface TooltipProps {
-  children?: string | React.ReactNode
-  enabled?: boolean
-  title?: string
-  shortcut?: string[]
-  tippyOptions?: Omit<Partial<Props>, 'content'>
-  content?: React.ReactNode
-}
-
-export interface TippyProps {
-  'data-placement': Placement
-  'data-reference-hidden'?: string
-  'data-escaped'?: string
-}
